requirements.txt
discord.py
gitpython
